// src/gen-ts/src/models/Sex.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 */
export const Sex = {
  M: 'M',
  F: 'F',
} as const
export type Sex = (typeof Sex)[keyof typeof Sex]

export function instanceOfSex(value: any): boolean {
  for (const key in Sex) {
    if (Object.prototype.hasOwnProperty.call(Sex, key)) {
      if (Sex[key as keyof typeof Sex] === value) {
        return true
      }
    }
  }
  return false
}

export function SexFromJSON(json: any): Sex {
  return SexFromJSONTyped(json, false)
}

export function SexFromJSONTyped(json: any, ignoreDiscriminator: boolean): Sex {
  return json as Sex
}

export function SexToJSON(value?: Sex | null): any {
  return value as any
}

export function SexToJSONTyped(value: any, ignoreDiscriminator: boolean): Sex {
  return value as Sex
}